Fix issue with import of code containing '<'
This fixes the problem that all code from an opening angle bracket
through to the next closing bracket or the end of the code element
is lost at import, since Nokogiri::XML::Node#replace interprets the
bracket as the start of an HTML tag.  By instead using the
content= method the new content string gets XML escaped.