Add link directly to ruby-core list
It is difficult to find the "Welcome to the "ruby-core" mailing list" email to manage your ruby-core list subscription. We should add a link directly to the page for easy access. I made the link body of to this site "subscribe" to keep it short, but we can pick something else. I think "subscribe" is not very accessible to people using screen readers, instead maybe "manage ruby-core list subscription".