Corrected a sentence and a broken link
About the french books link: it now leads to a website in japanese for some unknown reason. I've searched on Google for a list of books available to learn Ruby, and I found this list on developpez.com: http://ruby.developpez.com/livres/ which is a useful resource to learn programming in a lot of languages, so I thought this link might be useful. There are some duplicates with the english list, though. I cannot choose whether I put it in or not, so it's up to you. I removed it for now.