Add Rubyland aggregator to blog source list
Rubyland is an RSS/Atom feed aggregator of ruby news and blogs.
It's been up since November 2016, and is going well. It is the
only current ruby 'planet' style aggregator I know of. I
think it's a valuable source of ruby news and of finding
blogs.

Disclosure: I am the creator of rubyland.