Simplify 'check:links' task
The site should not have any broken links now, so do not distinguish
between "old" broken links (also existing on deployed site) and
"new" broken links (only existing for local build).
The previous behaviour was useful during the import of the site
from the old CMS.

Also do not abort when the first broken link is encountered;
the check runs for a very long time, and it's much better to
retrieve all broken links at a single run.