Partially revert "Remove link to digg programming page..."
This partially reverts commit b6bd84fb9c279e96ade8321e2016ee89bbfd5b88.
For `id' and `pl' the number of mentioned sites changes from plural
to singular, and I'm not convinced that the grammar is still correct.
Better an obviously outdated paragraph than a subtle error that is
likely to stay unnoticed.