Make other implementations more visible to users
The list of other implementations is hidden in the About page, which few developers will hit. By showing developers this list in the Downloads page, developers can be better informed and aware that the Ruby language is rich and widely adopted to the point of having multiple implementations and engines to run Ruby code.