Enable i18n for "Home" link in site navigation
Integrate "Home" link with the other sitelinks to allow for
localization via the `_data/locales/*.yml` files. For `en`:

```yaml
sitelinks:
- text: Home
  url: /en
  home: true
```

The `home` tag is introduced because the "Home" link requires special
treatment; it needs the CSS class "home" which controls the link's
visibility (it is only displayed in the hamburger menu for mobile view)
and ensures proper highlighting when selected.

As a side effect of this change the "Home" link is added to the sitelinks
in the footer, since the footer also uses `_includes/sitelinks.html`.
To keep the appearance of the site unchanged, the footer "Home" link
is set to `display: none`.