Do not specify teeny Ruby version in workflow
Do not specify a teeny version number for the Ruby to be used
but let the setup-ruby action use the latest available version.
To facilitate this, the CUSTOM_RUBY_VERSION environment variable
(that is needed to override the version specified in the Gemfile)
is set to the full MAJOR.MINOR.TEENY version string in a step that
uses the set-env action.