Fix syntax highlighted code blocks in lists (en)
Use fenced code blocks instead of Jekyll's `highlight` tag.

The latter does not work in list items: the highlight tag terminates
the list, so that the highlighted code is not part of the list item;
subsequent list items start a new list.

Visually, this fixes wrong indentation of the code blocks,
which had less indentation than the surrounding list items.

Fixes #2372.