Change search facility to DuckDuckGo
This commit replaces the Google search box to a DuckDuckGo search box.

I have chosen not to use their search box, but instead I have incorporated
this functionality into the search box that was previously being used.

I have also added short codes to every locale file, as seen in the previous
commit. DuckDuckGo supports localisation, so this should make the search
experience nicer for visitors of the Ruby website.

I am unfamiliar with Google, so this may be a guess, but I have witnessed
that some CSE special codes were missing from the locale files. I can only
guess that these were previously being used for localisation, e.g. choosing
a specific language for the search box.

If so, this should be a massive win for users of this site that are: not fluent
in the English language; do not have a Google account tailored to their language.

Moreover, this change will be a brilliant win for the privacy of the users of
your site. As Google is an advertising company, we can only guess what they do
with the data they collect on millions of users every day.

In my personal opinion, tackling this abhorrent abuse of data should also be
peformed by FOSS projects. Just think: every time someone searches, you've just
saved their online privacy! :D

Theoretically, DuckDuckGo should in fact be more performant than a Google search,
too. I have friends who have just confirmed this, which speaks volumes of
DuckDuckGo as a service.

I can't wait to see what happens!