Update references to freenode with references to Libera Chat
Freenode as the network we used to know is no longer. There's a new network now masquerading as the old.
See https://isfreenodedeadyet.com/.

Libera Chat is in many ways the spiritual successor run by the original maintainers of Freenode. The vast majority of the active community moved there. We from the [ruby-community.com](https://ruby-community.com/pages/irc_staff) project did so too. We've registered as a community project with Libera Chat and are happy to continue our services there. We're happy to providechannels and services to the Ruby community in #ruby and the #ruby- namespace on Libera Chat. Feel invited to talk to us in #ruby-community.

This pull request is making sure that users are not misdirected to a new and unmaintained environment but instead can find and join the community where it actually is.

* Update all references to #ruby on Freenode to #ruby on Libera Chat
* Lower entry barrier by linking to the webchat instead of using the irc:// protocol
* Dropped reference to #ruby-core, the channel seems inactive on both Libera Chat and "New Freenode"
* Dropped reference to #ruby-ja, the channel seems inactive on both Libera Chat and "New Freenode"
* Dropped reference to #ruby-tw, the channel seems inactive on both Libera Chat and "New Freenode"
* Dropped reference to #ruby-de, the channel seems inactive on both Libera Chat and "New Freenode"
* Dropped reference to #ruby.pl, the channel seems inactie on both Libera Chat and "New Freenode"
* Dropped reference to #ruby-id, the channel seems inactie on both Libera Chat and "New Freenode"