Added IBM text and link as a sponsor.
I don't think that we want to use IBM logo image for now. Because according to a
discussion with IBM folks, if we use the IBM logo image on our website, a
single individual in the Ruby project needs to sign IBM's trademark agreement
basically to make sure that the logo guidelines are followed.
The individual is either the trademark owner or the website owner.

Considering the cost and merit for us, I don't think we want to sign the
trademark agreement.